# openapi.yaml
openapi: 3.1.0
info:
  title: Trustify
  description: Software Supply-Chain Security API
  license:
    name: Apache License, Version 2.0
    identifier: Apache-2.0
  version: 0.6.0-rc.1
paths:
...
  /api/v3/team:
    get:
      tags:
      - tenancy
      summary: List teams
      operationId: listTeams
      parameters:
      - name: q
        in: query
        description: |
          EBNF grammar for the _q_ parameter:
          ```text
              q = ( values | filter ) { '&' q }
              values = value { '|', values }
              filter = field, operator, values
              operator = "=" | "!=" | "~" | "!~" | ">=" | ">" | "<=" | "<"
              value = (* any text but escape special characters with '\' *)
              field = (* must match an entity attribute name *)
          ```
          Any values in a _q_ will result in a case-insensitive "full
          text search", effectively producing an OR clause of LIKE
          clauses for every string-ish field in the resource being
          queried.

          Examples:
          - `foo` - any field containing 'foo'
          - `foo|bar` - any field containing either 'foo' OR 'bar'
          - `foo&bar` - some field contains 'foo' AND some field contains 'bar'

          A _filter_ may also be used to constrain the results. The
          filter's field name must correspond to one of the resource's
          attributes. If it doesn't, an error will be returned
          containing a list of the valid fields for that resource.

          An ASCII value of `NUL`, percent-encoded as `%00`, may be used
          to find resources on which a particular field isn't set. For
          example, `name=%00` and `name!=%00` yield the WHERE clauses,
          'NAME IS NULL' and 'NAME IS NOT NULL', respectively.

          Examples:
          - `name=foo` - entity's _name_ matches 'foo' exactly
          - `name~foo` - entity's _name_ contains 'foo', case-insensitive
          - `name~foo|bar` - entity's _name_ contains either 'foo' OR 'bar', case-insensitive
          - `name=` - entity's _name_ is the empty string, ''
          - `name=%00` - entity's _name_ isn't set
          - `published>3 days ago` - date values can be "human time"

          Multiple full text searches and/or filters should be
          '&'-delimited -- they are logically AND'd together.

          - `red hat|fedora&labels:type=cve|osv&published>last wednesday 17:00`

          Fields corresponding to JSON objects in the database may use a
          ':' to delimit the column name and the object key,
          e.g. `purl:qualifiers:type=pom`

          Any operator or special character, e.g. '|', '&', within a
          value should be escaped by prefixing it with a backslash.
        required: false
        schema:
          type: string
      - name: sort
        in: query
        description: |
          EBNF grammar for the _sort_ parameter:
          ```text
              sort = field [ ':', order ] { ',' sort }
              order = ( "asc" | "desc" )
              field = (* must match the name of entity's attributes *)
          ```
          The optional _order_ should be one of "asc" or "desc". If
          omitted, the order defaults to "asc".

          Each _field_ name must correspond to one of the columns of the
          table holding the entities being queried. Those corresponding
          to JSON objects in the database may use a ':' to delimit the
          column name and the object key,
          e.g. `purl:qualifiers:type:desc`
        required: false
        schema:
          type: string
      - name: offset
        in: query
        description: |-
          The first item to return, skipping all that come before it.

          NOTE: The order of items is defined by the API being called.
        required: false
        schema:
          type: integer
          format: int64
          minimum: 0
      - name: limit
        in: query
        description: |-
          The maximum number of entries to return.

          Zero means: return no items (the total count is still computed if requested).
        required: false
        schema:
          type: integer
          format: int64
          minimum: 0
      - name: total
        in: query
        description: Whether to compute and return the total count of matching items.
        required: false
        schema:
          type: boolean
      responses:
        '200':
          description: Matching teams
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResults_Team'
        '400':
          description: The request was not valid
        '401':
          description: The user was not authenticated
        '403':
          description: The user authenticated, but not authorized for this operation
    post:
      tags:
      - tenancy
      summary: Create a team
      operationId: createTeam
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamRequest'
        required: true
      responses:
        '201':
          description: The team was created
          headers:
            etag:
              schema:
                type: string
              description: Revision ID
            location:
              schema:
                type: string
              description: The relative URL to the created resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
        '400':
          description: The request was not valid
        '401':
          description: The user was not authenticated
        '403':
          description: The user authenticated, but not authorized for this operation
        '409':
          description: The external ID is already in use
  /api/v3/team/{key}:
    get:
      tags:
      - tenancy
      summary: Get a team
      operationId: getTeam
      parameters:
      - name: key
        in: path
        description: The ID of the team, or `ext:<external id>`
        required: true
        schema:
          type: string
      responses:
        '200':
          description: The team
          headers:
            etag:
              schema:
                type: string
              description: Revision ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
        '400':
          description: The request was not valid
        '401':
          description: The user was not authenticated
        '403':
          description: The user authenticated, but not authorized for this operation
        '404':
          description: The team was not found
    put:
      tags:
      - tenancy
      summary: Update a team, or create it when addressed by an external ID
      operationId: putTeam
      parameters:
      - name: key
        in: path
        description: The ID of the team, or `ext:<external id>`
        required: true
        schema:
          type: string
      - name: if-match
        in: header
        description: The revision to update
        required: false
        schema:
          type:
          - string
          - 'null'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamRequest'
        required: true
      responses:
        '200':
          description: The team was updated
          headers:
            etag:
              schema:
                type: string
              description: Revision ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
        '201':
          description: The team was addressed by external ID and has been created
          headers:
            etag:
              schema:
                type: string
              description: Revision ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
        '400':
          description: The request was not valid
        '401':
          description: The user was not authenticated
        '403':
          description: The user authenticated, but not authorized for this operation
        '404':
          description: The team was addressed by ID and was not found
        '409':
          description: The external ID is already in use
        '412':
          description: The provided If-Match revision did not match the actual revision
    delete:
      tags:
      - tenancy
      summary: Delete a team, including its memberships and role bindings
      operationId: deleteTeam
      parameters:
      - name: key
        in: path
        description: The ID of the team, or `ext:<external id>`
        required: true
        schema:
          type: string
      - name: if-match
        in: header
        description: The revision to delete
        required: false
        schema:
          type:
          - string
          - 'null'
      responses:
        '204':
          description: The team was deleted, or did not exist
        '400':
          description: The request was not valid
        '401':
          description: The user was not authenticated
        '403':
          description: The user authenticated, but not authorized for this operation
        '412':
          description: The provided If-Match revision did not match the actual revision
  /api/v3/team/{key}/member:
    get:
      tags:
      - tenancy
      summary: List the members of a team
      operationId: listTeamMembers
      parameters:
      - name: key
        in: path
        description: The ID of the team, or `ext:<external id>`
        required: true
        schema:
          type: string
      - name: offset
        in: query
        description: |-
          The first item to return, skipping all that come before it.

          NOTE: The order of items is defined by the API being called.
        required: false
        schema:
          type: integer
          format: int64
          minimum: 0
      - name: limit
        in: query
        description: |-
          The maximum number of entries to return.

          Zero means: return no items (the total count is still computed if requested).
        required: false
        schema:
          type: integer
          format: int64
          minimum: 0
      - name: total
        in: query
        description: Whether to compute and return the total count of matching items.
        required: false
        schema:
          type: boolean
      responses:
        '200':
          description: The members of the team
          headers:
            etag:
              schema:
                type: string
              description: Revision ID of the team
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResults_Member'
        '400':
          description: The request was not valid
        '401':
          description: The user was not authenticated
        '403':
          description: The user authenticated, but not authorized for this operation
        '404':
          description: The team was not found
    put:
      tags:
      - tenancy
      summary: Replace all members of a team
      operationId: setTeamMembers
      parameters:
      - name: key
        in: path
        description: The ID of the team, or `ext:<external id>`
        required: true
        schema:
          type: string
      - name: if-match
        in: header
        description: The revision of the team to update
        required: false
        schema:
          type:
          - string
          - 'null'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MembersRequest'
        required: true
      responses:
        '204':
          description: The members were replaced
          headers:
            etag:
              schema:
                type: string
              description: Revision ID of the team
        '400':
          description: The request was not valid
        '401':
          description: The user was not authenticated
        '403':
          description: The user authenticated, but not authorized for this operation
        '404':
          description: The team was not found
        '412':
          description: The provided If-Match revision did not match the actual revision
    patch:
      tags:
      - tenancy
      summary: Add and remove members of a team
      operationId: patchTeamMembers
      parameters:
      - name: key
        in: path
        description: The ID of the team, or `ext:<external id>`
        required: true
        schema:
          type: string
      - name: if-match
        in: header
        description: The revision of the team to update
        required: false
        schema:
          type:
          - string
          - 'null'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MembersPatch'
        required: true
      responses:
        '204':
          description: The members were updated
          headers:
            etag:
              schema:
                type: string
              description: Revision ID of the team
        '400':
          description: The request was not valid
        '401':
          description: The user was not authenticated
        '403':
          description: The user authenticated, but not authorized for this operation
        '404':
          description: The team was not found
        '412':
          description: The provided If-Match revision did not match the actual revision
...
components:
  schemas:
...
    MembersPatch:
      type: object
      description: Add and remove members of a team.
      properties:
        add:
          type: array
          items:
            type: string
          description: E-mail addresses of members to add. Unknown users will be created as invited.
        remove:
          type: array
          items:
            type: string
          description: E-mail addresses of members to remove.
    MembersRequest:
      type: object
      description: Replace all members of a team.
      required:
      - emails
      properties:
        emails:
          type: array
          items:
            type: string
          description: The e-mail addresses of the members. Unknown users will be created as invited.
...
    PaginatedResults_Member:
      type: object
      required:
      - items
      properties:
        items:
          type: array
          items:
            type: object
            description: A member of a team.
            required:
            - email
            - state
            properties:
              displayName:
                type:
                - string
                - 'null'
              email:
                type: string
              state:
                $ref: '#/components/schemas/UserState'
        total:
          type:
          - integer
          - 'null'
          format: int64
          minimum: 0
...
    PaginatedResults_Team:
      type: object
      required:
      - items
      properties:
        items:
          type: array
          items:
            type: object
            description: A team, a set of users, which can be granted roles as a unit.
            required:
            - id
            - name
            properties:
              description:
                type:
                - string
                - 'null'
              externalId:
                type:
                - string
                - 'null'
                description: |-
                  An optional ID, assigned by an external system.

                  The team can also be addressed using the key `ext:<external_id>`.
              id:
                type: string
                description: The ID of the team, assigned by the system.
              name:
                type: string
        total:
          type:
          - integer
          - 'null'
          format: int64
          minimum: 0
...
    Team:
      type: object
      description: A team, a set of users, which can be granted roles as a unit.
      required:
      - id
      - name
      properties:
        description:
          type:
          - string
          - 'null'
        externalId:
          type:
          - string
          - 'null'
          description: |-
            An optional ID, assigned by an external system.

            The team can also be addressed using the key `ext:<external_id>`.
        id:
          type: string
          description: The ID of the team, assigned by the system.
        name:
          type: string
    TeamRequest:
      type: object
      description: Mutable properties of a [`Team`].
      required:
      - name
      properties:
        description:
          type:
          - string
          - 'null'
        externalId:
          type:
          - string
          - 'null'
          description: An optional ID, assigned by an external system, unique across all teams.
        name:
          type: string
...
    UserState:
      type: string
      enum:
      - invited
      - active
      - disabled
...